Updating Dockerfile to fix running of parameter-binding application (#107)
Newly added instructions for getting /bin/ directory from a new application created
to run our own application -- due to some reason the application release/added on
the skf-labs doesn't have /bin/ directory in it -- which is required for the application
to work.

The following will create a new application with minimal config (required) and will then
move the /bin/ directory in our parameter-binding challenge directory and delete that
application later on. Solves the issue of running the application later on!

If this doesn't happen or /bin/ directory isn't created in the application (i.e. paramter-
binding), it asks us to create a new application and the error isn't verbose either